pyproject.toml
[project]
name = "webpilot-ai"
version = "0.1.0"
description = "AI-powered browser automation tool"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "your@email.com" },
]
dependencies = [
    "websockets",
    "playwright",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "openai",
    "google-generativeai",
    "langchain",
    "langchain-openai",
    "langchain-google-genai",
    "Pillow",
    "imageio",
    "imageio-ffmpeg",
    "SQLAlchemy",
    "psycopg2-binary",
    "langgraph",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
